+++ flutter_project/__brick__/lib/global/theme/theme.dart
@@ -3,15 +3,17 @@ import '../global.dart';
 
 import 'dark_theme.dart';
 import 'light_theme.dart';
-import 'theme_helper_service.dart';
+import 'theme_service.dart';
 
 export 'custom_theme_data.dart';
 
-final themeProvider = Provider<AppTheme>((ref) => AppTheme());
+final themeProvider = Provider<AppTheme>((ref) => AppTheme(ref.read));
 final themeModeProvider =
     StateProvider.autoDispose<ThemeMode>((ref) => ThemeMode.light);
 
-class AppTheme extends ThemeHelperService {
+class AppTheme extends ThemeService {
+  AppTheme(super.read);
+
   ThemeData get lightTheme => LightTheme.lightTheme;
   ThemeData get darkTheme => DarkTheme.darkTheme;
 }