+++ flutter_project/__brick__/lib/page/home/home_provider.dart
@@ -1,4 +1,5 @@
 import '../../global/global.dart';
 import 'controller/home_controller.dart';
 
-final homeProvider = Provider<HomeController>((ref) => HomeController(ref));
+final homeProvider =
+    Provider<HomeController>((ref) => HomeController(ref.read));