+++ test/hello/lib/global/theme/light_theme.dart
@@ -4,10 +4,33 @@ import 'package:google_fonts/google_fonts.dart';
 class LightTheme {
   static ThemeData get lightTheme => _lightTheme();
 
-  static _lightTheme() => ThemeData(
-        brightness: Brightness.light,
+  static _lightTheme([TextTheme? textTheme]) {
+    textTheme ??= ThemeData.light().textTheme;
+    const textStyleFunction = GoogleFonts.ubuntu;
+
+    return ThemeData(
 		useMaterial3:true,
-        primarySwatch: Colors.red, 
-        textTheme: GoogleFonts.ubuntuTextTheme(),
-      );
+        primarySwatch: Colors.green, 
+      brightness: Brightness.light,
+
+      // Text Theme
+      textTheme: TextTheme(
+        displayLarge: textStyleFunction(textStyle: textTheme.displayLarge),
+        displayMedium: textStyleFunction(textStyle: textTheme.displayMedium),
+        displaySmall: textStyleFunction(textStyle: textTheme.displaySmall),
+        headlineLarge: textStyleFunction(textStyle: textTheme.headlineLarge),
+        headlineMedium: textStyleFunction(textStyle: textTheme.headlineMedium),
+        headlineSmall: textStyleFunction(textStyle: textTheme.headlineSmall),
+        titleLarge: textStyleFunction(textStyle: textTheme.titleLarge),
+        titleMedium: textStyleFunction(textStyle: textTheme.titleMedium),
+        titleSmall: textStyleFunction(textStyle: textTheme.titleSmall),
+        bodyLarge: textStyleFunction(textStyle: textTheme.bodyLarge),
+        bodyMedium: textStyleFunction(textStyle: textTheme.bodyMedium),
+        bodySmall: textStyleFunction(textStyle: textTheme.bodySmall),
+        labelLarge: textStyleFunction(textStyle: textTheme.labelLarge),
+        labelMedium: textStyleFunction(textStyle: textTheme.labelMedium),
+        labelSmall: textStyleFunction(textStyle: textTheme.labelSmall),
+      ),
+    );
+  }
 }