+++ flutter_project/CHANGELOG.md
@@ -1,5 +1,6 @@
-# 0.1.1+3
+# 0.1.5-dev
 
 - Provider observer Logger Added
-- Change Default Theme to Light Mode from System Default
-- Set AppRouter routerProvider to GoRouter
+- ThemeHelperService name changed to ThemeService
+- TextTheme added with more control and DarkTheme text Color problem fixed.
+- isDarkMode method added to ThemeService using Provider and Context